Remove container/view IDs from insta snapshots (#8938)
Although these ID are supposed to be deterministic, they are in practice
not for reasons that are beyond me. Best eradicate them entirely from
insta snapshot.

To that end, I lost lots of time trying to use insta's redaction, before
realising I could just slap a `serde(skip)` in the data structures.

I also removed the `Deserialize` trait since that's not needed for
insta.

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>