Remove check for inner nullability of component columns (#9009)
### Related
* https://github.com/rerun-io/rerun/pull/8998
* #6819

### What
`pixi run rerun rerun://redap.rerun.io/catalog` failed with `Detected
malformed Chunk: The outer array in chunked component batch must be a
sparse list, got List(Field { name: "item", data_type: Utf8, nullable:
false, dict_id: 0, dict_is_ordered: false, metadata: {} })`.

The cause was a check in the `Chunk::sanity_check` requiring _inner
mutability_, i.e. allowing a single component to be null in a list. The
intention of the check was to check for _outer_ nullability, i.e. that
any cell in the column can be `null` (or a dense list).

In other words, we want to support `Vec<Option<Vec<T>>`, but NOT require
`Vec<Option<Vec<Option<T>>>`.

Why did this trigger now? Because we now allow sending component columns
as "mono components", i.e. as `Vec<T>` and that is then automatically
changed to `Vec<Vec<T>>`, but wether or not it has interior nullability
depends on if the source data had it.