Add general `re_sorbet::SorbetSchema` (#9017)
### Related
* Part of https://github.com/rerun-io/rerun/issues/8744

### What
Expand `re_sorbet` with `SorbetSchema` and `SorbetBatch` - a superset of
the existing `ChunkSchema` and `ChunkBatch`.

### How
I'm using a pretty ~ugly~ interesting approach of "fake inheritance":

Define `A` -> `B` as:
* `struct A` contains a `B`
* `impl Deref<Target=B> for A` (upcasting)
* `impl TryFrom<B> for A` (“downcasting”)
* `impl From<A> for B` (“object slicing”)

Using this notation:
* `ChunkSchema` -> `SorbetSchema`
* `ChunkBatch` -> `SorbetBatch` -> `arrow::RecordBatch`

This means the `ChunkBatch` can be used as `SorbetBatch` but also as an
`arrow::RecordBatch`.

Note that there is no dynamic dispatch involved at all.