gRPC spec update: Add support for fetching temporal range of recordings (#9010)
### What

Adding support for fetching chunks that belong to specific range of the
recording. This consists of the 2 APIs calls:
* fetching chunk ids for a given range
* fetching chunks for the given chunk ids (which we might want to
leverage separately in the future)

As there is a limitation for the client side gRPC streaming on the web
(see
[here](https://github.com/rerun-io/rerun/issues/8877#issuecomment-2652986044)),
we've decided to send a batch of chunk ids to the server. To make the
APIs somewhat symmetrical, the 1st API call will stream batches of chunk
ids.

Closes
* https://github.com/rerun-io/rerun/issues/8877