Add snapshot tests for time series view (#9036)
* Fixes https://github.com/rerun-io/rerun/issues/9031

Adds a small handful of tests to the time series.
Replaces the `check_scalar_clears.py` manual test. We have other manual
tests involving plots but they are all about overrides which I
considered out of scope for this round of testing for the very moment.

Upcoming TODO: Add multi-plot version for
* https://github.com/rerun-io/rerun/pull/9033