Add snapshot tests for multiple time series per entity (#9047)
Follow-up to both
* https://github.com/rerun-io/rerun/pull/9036
* https://github.com/rerun-io/rerun/pull/9033

Since the code paths for multiple time series (mostly due to
optimization) can vary quite a bit, I duplicated the previously added
tests to account for this.

Point series don't yet support multiple series on a single entity, so
all errors around them are expected - once we implement
https://github.com/rerun-io/rerun/issues/9020 these tests should then
automatically reflect this.


Fixed an issue when the number of names specified is smaller than the
number of series -- this sort of issue led me to adding a third variant
where we have a have a single property but multiple series. This ensure
that splatting where appropriate happens.