Document testing in Rerun (#8989)
Addition of many image comparison tests highlighted that we need to
document our tests.
Took the opportunity to list all test types we have that I could think
of and how to run time.

Furthermore, this now now serves as an faq for "but why does machine X
produce a different image than machine Y??" (A: it's a wonder they're as
close as they are ;-))