Switch back to the viewer when clicking a recording URI (#9064)
### Related

- Follow-up to #9018

### What

So far, clicking on a recording URI in the redap browser would silently
load the recording in the background. With this PR, the UI switches back
to viewer mode whenever a new recording is loaded, unless the user
cmd/ctrl- or middle-clicked the link, in which case the UI remains in
redap browser mode (mimicking the load tab in the background behaviour
of browser).

This involved a number of changes:
- `RecordingUri` is now removed, because the default `arrow_ui` handling
of links is sufficient.
- `rerun://` links are globally handled by the viewer, taking into
account the same-tab/new-tab behaviour in egui `OpenUrl` event.
- `ui.re_hyperlink` can now be specified is the new-tab behaviour is
forced or not. It is the case for our documentation link, but not for
generic urls (e.g. from `arrow_ui()`
- Some cleanup around `redap::Scheme` happened.