Identify timelines uniquely by name (ignore type) (#9097)
### Related
* Part of https://github.com/rerun-io/rerun/issues/9084
* Sibling PR: https://github.com/rerun-io/dataplatform/pull/287

### What
We used to uniquely identify a time column by its name and _type_
(temporal or sequence).

This PR changes this to only use the name.

Using the same name but different types has always been undefined
behavior, and a bad idea. If you've been doing this, it is likely that
your old code already produced "surprising" results. After this PR, you
should at least get a clear warning printed in the terminal.

### What changed
* Store
* Caches
* Queries
* Dataframe column selectors

### What happens when you mix different time types on the same timeline?
Both the SDK and the viewer will log a warning, and the last used
`TimeType` will be used in the UI.

### TODO:
* [x] Detect switching the type of time column _in the SDK_, and log a
warning.
* [x] Full check
* [x] Migration guide entry
* [x] Update docs where relevant
* [x] Dataplatform PR

### Future PR
* `TimePoint`