Fix support for numpy-2  (#9109)
### What
- Fixes https://github.com/rerun-io/rerun/issues/9105
- Fixes https://github.com/rerun-io/rerun/issues/8996
- Fixes https://github.com/rerun-io/rerun/issues/8486
- Fixes https://github.com/rerun-io/rerun/issues/6762
- Fixes https://github.com/rerun-io/rerun/issues/4807


In order to move our pixi project to numpy to, we also had to upgrade
pytorch to 2.3+ (we might consider jumping all the way to 2.6, but I was
trying to find the minimum bump to risk breaking other things).

Unfortunately torch-2.2 is the last version that supported intel macs.

I disabled the wheel-testing on nightly with an issue to re-enable:
https://github.com/rerun-io/rerun/issues/9108

<!--
Make sure the PR title and labels are set to maximize their usefulness
for the CHANGELOG,
and our `git log`.

If you have noticed any breaking changes, include them in the migration
guide.

We track various metrics at <https://build.rerun.io>.

For maintainers:
* To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.
* To deploy documentation changes immediately after merging this PR, add
the `deploy docs` label.
-->