Test that we don't break backwards compatibility of .rrd files (#9133)
### Related
* Part of #9110

### What
Adds a bunch of `.rrd`:s to `git lfs`, that we should keep there as a
test of backwards compatibility.
If any new code breaks loading of these old .rrd files, the CI will
complain.

Running locally:
> pixi run check-backwards-compatibility

### What if the CI complains about my PR?
Then your code broke backwards compatibility. Can you make the change so
that it doesn't? If not, consult with me! We are not yet promising
backwards compatibility, but we should at least make a reasonable
effort.

### What does this PR cover?
All Rust snippets are tested, as are the main examples (as of today).
This should include most components, but does not cover _everything_. So
even if `pixi run check-backwards-compatibility` passes, it is possible
that we have broken backwards compatibility in some subtle way. But this
is at least a start.

### TODO
* [x] Add a single pixi command to verify the files

### Future work
* Add more files, for improved coverage
* Automatically detect if some snippets/examples are missing from the
test dataset

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>