Editable time control & time context menus (#9174)
### Related

* Fixes [#7653](https://github.com/rerun-io/rerun/issues/7653)


### What


https://github.com/user-attachments/assets/9ea1e396-6ad7-4a27-9f5e-648903eafaee

Time in the timeline panel is now an input field that supports
* raw timestamps
* relative times
* date times

Edits are applied when the field looses its focus (you can also just
press `enter` for that).

Furthermore, there's now context menus to copy timeline name & raw
timestamps now.

Previously, the time panel was dragable in some situations (time indices
not exceeding a certain numerical range). We could bring that back, but
I decided to drop it in favor of a simple edit field - dragging would
have made this a lot more complicated.