Deprecate `rr.new_recording()` in favour of `rr.RecordingStream()` and improve type checking of the Rust bindings (#9206)
### Related

* Part of https://github.com/rerun-io/rerun/issues/9187

### What

The only way to recreate a `RecordingStream` so far was with
`rr.new_recording()`. In particular, the `rr.RecordingStream()`
constructor could not be used by end users. This is an anti-pattern and
against the idea of a first-class stateless API.

This PR:
- deprecates `rr.new_recording()`...
- ...in favour of a well-behaved `RecordingStream` constructor

And in order to achieve this with the help of mypy, this PR also:
- adds `rerun_bindings` to mypy's coverage
- adds all missing `rerun_bindings` stubs (thanks claude)
- fixes several errors in the bindings
- fixes annotations in tests 

Finally (?), to make `python_check_signatures`, `py-fmt`, and `py-lint`
all go along with each other, this PR:
- add docstrings to all symbols exported by the `rerun_bindings`
- adjust existing docstrings to make the linter happy
- copies all docstrings to the stubs

---------

Co-authored-by: Andreas Reich <andreas@rerun.io>