Refactor time code, and change formatting of time ranges in URIs (#9241)
### Related
* Follows https://github.com/rerun-io/rerun/pull/9236
* Part of #8635 

### What
A lot of refactors and cleanups in preparation for the above issue.

`struct Time` is going away (but is not yet gone).
In its place will be `struct Duration` and the new `struct Timestamp`.
I've started moving code out of `Time` and onto these types.
I've also started using [`jiff`](https://crates.io/crates/jiff), which
is looking to become the best time/date library.

`IndexCell` now implements `FromStr` and `Display`, using jiff. This is
used for the time range parameters for the gRPC URI.
 The format is one of:

* `123` for sequences
* `123.123s` for durations
* `2022-01-01T00:00:03.123456789Z` for timestamps

The duration and timestamps should follow ISO 8601

### Do-no-merge
Because the merge target is
* https://github.com/rerun-io/rerun/pull/9236

---------

Co-authored-by: Andreas Reich <r_andreas2@web.de>