Add entity search to the add/remove entity modal (#9120)
### What

Add entity search to the add/remove entity modal, similar to the one in
the time panel. This PR introduces an alternative UI for the search
filter that is a better fit for the add/remove entity modal

Note: `EntityData` is heavily copy/pasted from the time panel's version.
It sort of feels it should be deduplicated, but (1) I'm not sure where
to put it and (2) there are some customisation that would be annoying to
abstract away. So I'm actually fine leaving it as is.


https://github.com/user-attachments/assets/4a4247d3-6617-4bc7-8942-9cbcb7e958c9

---------

Co-authored-by: Lucas Meurer <lucasmeurer96@gmail.com>