feat: update arrow to 54 (#9259)
### Related

* Part of https://github.com/rerun-io/rerun/issues/3741

### What

Now that lance, lancedb, and datafusion have all updated to arrow-54,
this PR is aimed to upgrade rerun. The arrow update includes an update
to pyo3, which deprecates some features and adds in additional
requirements about having all python classes be `Sync`, though this last
point does not seem to impact our repo.