Prevent automatic opening of views for archetypes in `RecordingProperties` (#9315)
### Related

* Closes #9313

### What

This PR adds `- /__partition_properties/**` to the filter list unless
overwritten. This unifies our dataframe APIs again and is a first step
to logging arbitrary data to properties.

### Test

To test, it's probably best to look at one of the
`recording_properties.{cpp|py|rs}` snippets of the follow up PR:

* #9316

Here is a demo of the UI interactions:


https://github.com/user-attachments/assets/6158b77b-6200-453c-b5e8-80b9a854d25c