manifest registry service spec update: make partition id a string type (#9324)
### What

Partition ids can be user defined fields (episode id, recording id,
etc). For simplicity we now represents in the manifest registry service
with a simple string type. What actual field and type defines the
partition id will most likely be defined as part of creating the
``Dataset``.

This will also simplify the initial integration with ``rrd`` and the
``StoreId::id()``.

We might revisit how we represent and deal with the partition ids e2e,
but it will be simpler to do that as concrete cases and client
integration comes along.