Provide APIs to log arbitrary data to recording properties (#9316)
### Related

* Closes: #9199 

### What

This PR should cover all use-cases that we could come up with so far and
be flexible enough to extend into the future without breaking changes.
Here is the gist:
* We map a property name to an `EntityPath` that lives at
`/__partition_properties/<property-name>`. Builtin recording properties
(name, start time) will be at the special path
`/__partition_properties/recording`.
  * Nothing lives directly at the root.
  * The current `RecordingProperties` archetype stays as is.
* The value of a property is always an archetype (potentially
AnyValues). In the future we might open this up to specify general Arrow
values.
* The API will be: `rec.send_property(name: str, archetype: AsComponent
| [AsComponent])`
* Motivation: `send_` signals that the properties are not guaranteed in
a distributed setting.
* `name` is positional, so we can upgrade it to reflect hierarchies
later without breaking user code. (For example by renaming to
`entity_path`).
* For now we don't expose the fact that this modeling allows for
hierarchical grouping of properties.
* We will have to special methods to set the builtin properties:
  * `send_recording_name`
  * `send_recording_start_time_nanos`
* When creating a `RecordingStream` with an explicit recording ID, we
don't send the builtin properties automatically, because that means we
are usually in "append mode".

Open questions:
* The selection panel currently expects component-level data. Can we
share the work for implementing something like re_archetype_ui with the
catalog group so we could use that both in the tables and in the
selection panel?

---------

Co-authored-by: Andreas Reich <andreas@rerun.io>