Prevent modals from shrinking / moving around when the content changes (#9307)
### What

This prevents modals from shrinking if their content size changes, which
is annoying when the size rapidly changes, e.g. during autocomplete.

The size will be reset if the screen size changes, so that the modal can
adjust accordingly.

Question is wether the prevent_shrinking call should happen inside or
outside the frame.

Outside:


https://github.com/user-attachments/assets/36c65743-1968-4a28-a68b-10e72ec73ba9

Inside:


https://github.com/user-attachments/assets/028b2230-7fa2-4561-a6bd-bd2ac8a5d579

As is, the PR implements the Outside behavior (which I think is nicer),
switching it would be simple though.

@gavrelina which one do you prefer?