Initial python wrapper for the new catalog API (#9301)
### Related

- follow-up to #9299
- part of https://github.com/rerun-io/dataplatform/issues/400
- part of [#9360](https://github.com/rerun-io/rerun/issues/9360)

### What

This PR introduces the new catalog client API to the Python SDK. This is
very much WIP, with the on-going todo list being captured in #9360. This
PR primarily introduce some core patterns:

- How the connection is managed though a `ConnectionHandle`, which is
owned by the root `CatalogClient` instance but cloneable at will by
related instances and method implementation.
- An `Entry` hierarchy, with a `Dataset` sub-class, to model the object
found in the catalog.
- Error management.

---------

Co-authored-by: Tim Saucer <timsaucer@gmail.com>