Explicit control over sending properties (#9356)
### What

Previously, we did not send the initial recording properties chunk when
`recording_id` was set on `RecordingStreamBuilder`. This weird
interactions with the global Python SDK though.

This PR reverts that change and instead provides better control over
when properties are send.

It also fixes a typo in the `recording_properties` snippet that caused
`.rrd`-compare failures in CI.

---------

Co-authored-by: Antoine Beyeler <49431240+abey79@users.noreply.github.com>