Improve snippet index (#9397)
### Related
Improves the text matching to be "whole word" so that the `Image`
archetype is not a match when the snippet is actually only using the
`DepthImage` archetype.

This came up because I wanted to forbid references to deprecated
objects, and realized we got a lot of false positives (e.g. matching
deprecated `Scalar` when using the new `Scalars` (notice the `s`)).

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>