Don't trigger callbacks on welcome screen (#9409)
Right now when the Viewer is "empty" (=no active recordings), it still
has an active app, which is the welcome screen, and that has an empty
recording in it. As a result, we trigger time update callbacks, even if
the user doesn't see any active recordings. This is not only confusing
but also actively harmful behavior for anyone who's trying to use
callbacks to build on top of the Viewer.

The solution here is to disable timeline callbacks if we don't have an
active recording. We define that as "has an app id which isn't the
welcome screen app id".

Additionally, we don't trigger timeline callbacks if the active timeline
is empty (=has no timepoints).