Use multi-dimensional scalar values when loading LeRobot dataset (#9402)
### Related

* Closes #9318 

### What

We now use the multi-scalar support added in #9005 when loading
multi-dimensional action/observation values from a LeRobot dataset.
This is the preferred as it aligns with how data is consumed during
training, each frame’s actions or observations are expected as a single
tensor rather than a collection of separate scalar values. Storing it as
one neat array makes it easier to query action/observation values back
out of Rerun, as there's now just a single entity path.