Update redap browser to use the new catalog API (#9413)
Title. This implementation is somewhat quick-and-dirty: the _entire_
list of datasets and their partition table is downloaded at once,
without per-dataset granularity. This will have to be improved in the
future, but that at least unblocks improving the UI and removing the
legacy API.

*Note*: The new url scheme used in the dataset table is not yet
supported, so it breaks the ability to load partition to the viewer.
(I'll work on that next.)