Fix optional notebook package import (#9415)
`rerun_notebook` is an optional dependency. Right now, we attempt to
import `rerun_notebook` unconditionally, and ignore the `ImportError` if
there is one. It turns out this is not enough, because the
`rerun_notebook` package is attempting to load a file which does not
exist, which raises `FileNotFoundError` instead.

We can assume the widget always exists if the package is imported, so I
updated the logic to actually check if the package is installed, and
only attempt to import it if so.

Testing:
```
rm -rf .pixi/envs/examples
pixi run py-build-examples
pixi run -e examples minimal
```
Running this on main results in a `FileNotFoundError`, which is gone on
this PR.