Introduce URI scheme for dataset and add support for opening remote recordings using the new API (#9422)
Introduce the URL scheme for downloading dataset data:

```
$ORIGIN/dataset/$DATASET_ID/data?partition_id=$PARTITION_ID[&time_range=$TIME_RANGE]
```

Relevant slack thread for the backing discussion:
https://rerunio.slack.com/archives/C08EWSCS26P/p1743092658188489

For now, we only support fetching a single partition, so the
`partition_id` parameter is mandatory. In the future, we will support
arbitrary queries. The `time_range` parameter is optional. If used, the
time range will be highlighted, but filtering on actual time range when
fetching chunk is not yet supported
(https://github.com/rerun-io/dataplatform/issues/474).

NOTES:
- Review commit-by-commit. The first commit is a pure move refactor.
- I've updated the existing `Recording` tests instead of adding new
ones, because this stuff will soon be removed anyways.