Introduce `attr.docs.state` to our codegen (#9431)
### Related
* Part of https://github.com/rerun-io/rerun/issues/9427

### What
To make it easier to manage the "stabilization" state of our types, I've
introduced a new attribute (that replaces several old ones).

My plan is to make this mandatory for all archetypes and components,
with the benefits:
A) force us to consider its state when adding a new type
B) make sure we don't forget to set it to `unreleased` (to add
`?speculative-link`)

I will make it mandatory in a future PR though, to keep this PR smaller.

### Docs for `attribute "attr.docs.state";`
Must be set on all archetypes and components.

Must have one of the following values:
- `unreleased`
- `experimental`
- `unstable`
- `stable`
- `deprecated`

### `unreleased`
Any external links referencing this datatype will be given a speculative
link marker.
Speculative links contain a `?speculative-link` query param. Any such
links are ignored by linkinator,
and we check for their absence as the first step in our release process.

### `experimental`
New API that is not yet fully supported, and may change significantly in
future versions.
Implies everything that `unstable` implies.

### `unstable`
Used for types that are likely to be removed or changed significantly,
and in a way that the data won't be backwards compatible.

### `stable`
Used for types that are unlikely to be removed or changed significantly.
If they are changed, we will make sure that the old data can still be
read.

### `deprecated`
Marks something as deprecated followed by a (mandatory!) migration note.

If specified on an object (struct/enum/union), it becomes deprecated
such
that using the object should emit a warning in all target languages.
Furthermore, documentation will mention that the object is deprecated
and display
the specified migration note.

If set, you MUST also set `attr.docs.deprecated_since` and
`attr.rerun.deprecated_notice` to a non-empty string.