minor: add python dependencies for examples (#9250)
### What

This PR adds two python dependencies for use in the examples.
Additionally, it moves the `remote` feature to be used by default.

I chose not to remove `remote` as a feature and just include its
dependencies so that users who are extending rerun without remote
connection capability can choose to opt out of it.