Support `#datapath` fragment on redap URL (#9478)
This adds support for appending a data path (e.g. `/some/entity[#42]`)
to a redap URL, as a `#fragment`.

When clicking a redap URI with such a fragment, we will focus on that
entity.
This is untested, however, and depends on us keeping the focus until the
entity is loaded.

A `DataPath` is currently defined as

```rs
pub struct DataPath {
    pub entity_path: EntityPath,
    pub instance: Option<Instance>,
    pub component_name: Option<ComponentName>,
}
```

We will wanna extend this in the future to support tagging, and maybe
even a timepoint.

### Extending it with time point

It's not clear how to specify _both_ time _and_ a data path in the
fragment 🤔 I guess we need to use a special character, like comma:

`#/entity/path,log_tick=42` maybe

…and then comes the trouble of escaping

---------

Co-authored-by: Andreas Reich <andreas@rerun.io>