redap manifest registry: add endpoints for asynchronous dataset/partitions registration  (#9487)
### Related
* [dataplatform/#528](https://github.com/rerun-io/dataplatform/pull/528)

### What
Introduces a new `ManifestRegistry` endpoint called
`RegisterWithDatasetAsync`, which - surprise- is the async version of
`RegisterWithDataset`.

It also introduces new methods for the `Tasks` service to query the
output of a task once completed.

**Recommended reviewing commit-by-commit**

<!--
Make sure the PR title and labels are set to maximize their usefulness
for the CHANGELOG,
and our `git log`.

If you have noticed any breaking changes, include them in the migration
guide.

We track various metrics at <https://build.rerun.io>.

For maintainers:
* To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.
* To deploy documentation changes immediately after merging this PR, add
the `deploy docs` label.
-->

---------

Signed-off-by: Andrea Reale <andrea@rerun.io>