Format random IDs withour dashes (#9582)
### Related
* Part of https://github.com/rerun-io/dataplatform/issues/250

### What
We sometimes use a string-formatted UUIDv4 as a random ID. I think it's
better we skip the dashes in these cases, so that we can
double-click-to-select.

Before: `0840159d-1af9-48fb-8d14-208e388b26ee`
After: `0840159d1af948fb8d14208e388b26ee`

For AppId I also decided to prefix the id with `app_`:
`app_0840159d1af948fb8d14208e388b26ee`

It also makes it clear that this is indeed not _always_ a UUID. For
instance, our AppId is a custom string, but _somtimes_ a random string
(currently generated with a UUID, but that isn't fixed).