rerun_py: shove support for async register in `register` dataset
We replaced the synchronous register_with_dataset endpoint with an
asynchronous equivalent. The python client, however, expects the
endpoint to be synchronous.

This change is a quick-fix to make the python client block for dataset
registration task completion so that current user should not feel the
difference.

We should improve this in the future by exposing the asynchronous nature
of the endpoint directly to the client.

Signed-off-by: Andrea Reale <andrea@rerun.io>