Remove all uses of `arrow` and `arrow2` from `re_types_builder` (#9724)
### Related
* Part of https://github.com/rerun-io/rerun/issues/3741

### What
This replaces the use of `arrow` and `arrow2` in `re_types_builder` with
more explicit enums.

In order to maximize the chance of success (and minimize the diff) I
decided to mostly follow the old structure of the code, including
maintaing `.to_logical_type()`.

`arrow2/re_arrow2` is now GONE, but there are some minor cleanups I
wanna do before fully closing #3741