rerun_py: adapt the `register` API to the new asynchronous nature of  `register_with_dataset` (#9590)
### Related

* Related to: https://github.com/rerun-io/dataplatform/issues/527
* Follows:
  * https://github.com/rerun-io/rerun/pull/9537
  * https://github.com/rerun-io/dataplatform/pull/543 

### What
In https://github.com/rerun-io/dataplatform/pull/543, we replaced the
synchronous register_with_dataset endpoint with an
asynchronous equivalent. The python client, however, expects the
endpoint to be synchronous.

This change is a quick-fix to make the python client block for dataset
registration task completion so that current user should not feel the
difference.

We should improve this in the future by exposing the asynchronous nature
of the endpoint directly to the client.

Signed-off-by: Andrea Reale <andrea@rerun.io>