Do not hide all entities when the search is active but empty (#9734)
### Related

* closes https://github.com/rerun-io/rerun/issues/9520

### What

This PR changes the behaviours of the entity search filter when the
search is activated but the query is still empty. Now, the displayed
content is unchanged compared to when the search is inactive.
Previously, nothing would be displayed in that situation. This required
subtle change of when a session id is returned, such that the collapse
state used in the active-but-empty-query state is the same as in the
inactive state.