Compare snippets to legacy rrds (#9751)
## Related
* Requires https://github.com/rerun-io/rerun/pull/9750
* Part of #9110

## What
This restructures the folder of .rrds that we use for
backwards-compatibility checks, so that it has a `snippets` folder,
matching the structure of our snippets. These are then read by
`compare_snippets_output.py` and are compared to the latest output of
the snippets. They should be equal to each other (since the old .rrds
should be migrated on ingestion).

### Consequences
If we add a new snippet, the CI will fail until the output .rrd of that
snippet is checked in to CI (using `compare_snippets_output.py
--write-missing-backward-assets`)

If we _change_ a snippet (so that it outputs something different), we
need to either:
* Update the file on `git lfs`
* Add a _new_ file to `git lfs` alongside the old one
* Add an opt-out flag for comparing to the file on `git lfs` to
`snippets.toml`

Since this should be a rather rare occurrence (our snippets tend to stay
pretty stable), any one solution will do, and I have not decided on one
yet. Maybe cross that bridge if and when we get to it?