Introduce a DataFusion-backed table widget (#9764)
### Related

* part of https://github.com/rerun-io/rerun/issues/9741

### What

This PR introduces `DataFusionTableWidget`, a widget capable of
displaying a table in a datafusion session context, along with the
necessary changes to use it for partition table display in the redap
browser.

Notes:
- The core piece is the `DataFusionAdapter`, which triggers and track
completion of datafusion queries (which are async), and uses egui memory
to save state across frames.
- A `TableBlueprint` stuff makes an appearance. This eventually leads to
the discussion of manipulating them from code, etc.
- The pre-existing `TableConfig` remains. At first sight, it seems that
it should be merged with `DataFusionAdapter`, since both construct
implement manipulation of the dataframe being displayed. There are good
reason to keep them separate though:
- It makes sense to keep column visibility/order at the RecordBatch/UI
level and outside of the (slower) datafusion loop.
- Eventually, I would like that logic to be implemented directly into
`egui_table`, since it is potentially useful for any table user.
- Likewise, there is a small `Columns` helper. This one will definitely
be merged with `TableConfig` as soon as the other use of `TableConfig`
are replaced by `DataFusionTableWidget` (namely: the in-viewer table).