Improve and mitigate warnings around dataloss when flushing (#9846)
### Related

* Closes #9818.

### What

> [!IMPORTANT]
> This PR also changes the way `RecordingStream` is free'd in the C/C++
API. Before we called `stream.disconnect`, which unnecessarily replaced
the current sink with a _buffered_ sink that would be immediately
dropped afterwards. Not only did this cause spam in the log outputs, it
also lead to race conditions upon (log) application shutdown.

This PR makes it more explicit why we drop data during flushing, by
bumping the log messages to `warn!`. It also improves the message by
pointing the users to `flush_timeout`.

We also bump the default timeout from two seconds to now 3 seconds.

It's worth taking note that explicitly calling `flush_blocking` from our
SDKs should be able to opt-out of this timeout, to ensure all data is
sent. This will be tracked here:

* #9845.