redap: support REDAP_TOKEN env var
To access redap endpoints that require authentication (such as
`sandbox.redap.rerun.io`) we need ways to pass the auth token to the
GRPC client.

This is the simplest most naive way, that unblocks us.

We may consider to have explicit SDK params and even UI elements in the
viewer for that.

Signed-off-by: Andrea Reale <andrea@rerun.io>